+++ pkg/http/api_validators.go
@@ -59,3 +59,15 @@ func validatePrefixQuery(value string) (string, error) {
 	}
 	return value, nil
 }
+
+// Helper: Validate neighbors query. A valid query should have
+// at least 4 chars.
+func validateNeighborsQuery(value string) (string, error) {
+	if len(value) < 3 {
+		// Maybe make configurable,
+		// A length of 3 would be sufficient for "DFN" and
+		// other shorthands.
+		return "", ErrQueryTooShort
+	}
+	return value, nil
+}